+++ webarchive/webarchive.py
@@ -68,12 +68,12 @@ class WebArchive(object):
                 archive = plistlib.load(fp)
 
         # Process the main resource
-        self._main_resource = WebResource(archive["WebMainResource"])
+        self._main_resource = WebResource(self, archive["WebMainResource"])
 
         # Process subresources
         if "WebSubresources" in archive:
             for plist_data in archive["WebSubresources"]:
-                res = WebResource(plist_data)
+                res = WebResource(self, plist_data)
                 self._subresources.append(res)
 
                 # Make local paths for extracting each subresource